requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-generativeai==0.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.9.0
mangum==0.17.0